Static cast

+++ src/pso/bench/outputPieceSets.cpp
@@ -5,9 +5,9 @@ int main() {
   PieceSetGetter ps;
   auto sets = ps.getPieceSets(100);
   for (const auto &game: sets) {
-    for (int i = 0; i < game.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(game.size()); ++i) {
       std::cout << game[i];
-      if (i != game.size()-1) std::cout << ' ';
+      if (i != static_cast<int>(game.size())-1) std::cout << ' ';
     }
     std::cout << '\n';
   }
